requirements.txt
Flask==3.0.0
flask-cors==4.0.0
pymongo==4.6.1
PyJWT==2.8.0
Werkzeug==3.0.1
dnspython==2.6.1
